arreglado bug de clonacion de promociones y coeficientes

+++ db/migrate/20150610135929_create_values.rb
@@ -2,7 +2,7 @@ class CreateValues < ActiveRecord::Migration
   def change
     create_table :values do |t|
       t.string :quota
-      t.decimal :value
+      t.float :value
       t.belongs_to :coefficient, index: true
       t.timestamps
     end
